+++ .gitignore
@@ -1,1 +1,3 @@
 .vscode/*
+.vs/
+.git/
