+++ data/systems.csv
@@ -1,0 +1,6 @@
+name,coordinateX,coordinateY,coordinateZ
+Stanton,49.534718,-2.633965,16.475292
+Pyro,40.126133,8.414866,-17.217419
+Nyx,26.538177,27.071026,-14.517497
+Terra,67.713928,-8.1313095,-3.1416104
+Magnus,49.620834,-13.12486,2.547168
